requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
tortoise-orm==0.25.0
aiosqlite==0.19.0  # SQLite支持（开发环境）
asyncpg==0.29.0   # PostgreSQL支持（生产环境）